requirements.txt
discord.py>=2.4
aiohttp
environs
asyncpg
redis
